Move indexation error to dedicated file
